+++ client/projects/client/src/app/modules/chat/chat.module.ts
@@ -1,27 +1,28 @@
-import {CommonModule} from "@angular/common";
-import {provideHttpClient, withInterceptorsFromDi} from "@angular/common/http";
-import {NgModule} from "@angular/core";
-import {ReactiveFormsModule} from "@angular/forms";
-import {UiModule} from "@ui";
-import {XmppEventsDistributionService} from "../../lib/services/xmpp-events-distribution.service";
-import {MainChatComponent} from "./main-chat.component";
-import {ChatRoutingModule} from "./chat-routing.module";
-import {SettingsComponent} from "./containers/settings/settings.component";
-import {TabFormComponent} from "./components/tab-form.component";
-import {ConversationComponent} from "./containers/chat/conversation.component";
-import {ConversationsListComponent} from "./components/conversations-list/conversations-list.component";
-import {ContactsComponent} from "./containers/contacts/contacts.component";
-import {HeaderComponent} from "../../layout/header/header.component";
-import {SidebarComponent} from '../../layout/sidebar/sidebar.component';
-import {ConvItemComponent} from "./components/conv-item.component";
-import {ConversationAreaComponent} from "./components/conversation-area/conversation-area.component";
-import {ChatMenuComponent} from "./components/conversation-area/chat-menu.component";
-import {ChatStateComponent} from "./components/chat-states.component";
-import {
-    CollocutorItemComponent
-} from "./components/collocutor-item.component";
-import {MessageComponent} from "./components/conversation-area/message.component";
+import { CommonModule } from "@angular/common";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
+import { NgModule } from "@angular/core";
+import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { UiModule } from "@ui";
+import { XmppEventsDistributionService } from "../../lib/services/xmpp-events-distribution.service";
+import { MainChatComponent } from "./main-chat.component";
+import { ChatRoutingModule } from "./chat-routing.module";
+import { SettingsComponent } from "./containers/settings/settings.component";
+import { TabFormComponent } from "./components/tab-form.component";
+import { ConversationComponent } from "./containers/chat/conversation.component";
+import { ConversationsListComponent } from "./components/conversations-list/conversations-list.component";
+import { ContactsComponent } from "./containers/contacts/contacts.component";
+import { HeaderComponent } from "../../layout/header/header.component";
+import { SidebarComponent } from '../../layout/sidebar/sidebar.component';
+import { ChatRoomItemComponent } from "./components/chat-room-item.component";
+import { ConversationAreaComponent } from "./components/conversation-area/conversation-area.component";
+import { ChatMenuComponent } from "./components/conversation-area/chat-menu/chat-menu.component";
+import { ChatStateComponent } from "./components/chat-states.component";
+import { ParticipantItemComponent } from "./components/conversation-area/chat-menu/participant-item.component";
+import { MessageComponent } from "./components/conversation-area/message.component";
 import { SendFromComponent } from "./components/conversation-area/send-form.component";
+import { NoChatSelectedComponent } from "./components/not-chat-selected.component";
+import { LibPipesModule } from "@lib";
+import { PrivateRoomItem } from "./components/conversation-area/chat-menu/private-room-item.component";
 
 @NgModule({
     declarations: [
@@ -35,19 +36,24 @@ import { SendFromComponent } from "./components/conversation-area/send-form.comp
         // ui
         TabFormComponent,
         ConversationsListComponent,
-        ConvItemComponent,
+        ChatRoomItemComponent,
         ConversationAreaComponent,
         ChatMenuComponent,
         ChatStateComponent,
-        CollocutorItemComponent,
+        ParticipantItemComponent,
         MessageComponent,
         SendFromComponent,
+        NoChatSelectedComponent,
+        PrivateRoomItem,
     ],
     imports: [
         CommonModule,
+        FormsModule,
         ReactiveFormsModule,
         ChatRoutingModule,
         UiModule,
+
+        LibPipesModule,
     ],
     providers: [
         provideHttpClient(withInterceptorsFromDi()),