+++ client/projects/client/src/app/modules/chat/components/conversation-area/send-form.component.ts
@@ -4,7 +4,9 @@ import { FormControl, FormGroup, Validators } from "@angular/forms";
 @Component({
     selector: "app-send-form",
     template: `
-    <form [formGroup]="form" (ngSubmit)="onSubmit()" class="border px-5 py-4 flex items-center gap-4 m-4 rounded-xl">
+    <form [formGroup]="form" (ngSubmit)="onSubmit()"
+        class="border px-5 py-4 flex items-center gap-4 m-4 rounded-xl"
+    >
         <div class="flex items-center gap-4 text-slate-300">
             <button type="button" disabled><i class="fa-solid fa-microphone"></i></button>
             <button type="button" disabled><i class="fa-regular fa-image"></i></button>
@@ -36,7 +38,6 @@ export class SendFromComponent {
         if (!this.form.valid) {
             return
         }
-        console.log(this.form.value);
         this.form.reset();
     }
 }