+++ client/projects/client/src/app/modules/chat/components/conversations-list/conversations-list.component.ts
@@ -1,9 +1,11 @@
 import { Component } from "@angular/core";
-import { UiConv, mock_conversations } from "@lib";
+import { Room } from "@lib";
 import { Store } from "@ngrx/store";
-import { map, merge, Observable, of, scan } from "rxjs";
-import { AppState } from "../../../../store/store";
-import { selectRoomsAll } from "../../../../store/chat/reducers/rooms";
+import { BehaviorSubject, combineLatest, map, Observable } from "rxjs";
+import { AppState } from "@store/store";
+
+import { Router } from "@angular/router";
+import { selectRoomsAll } from "@store/chat/selectors/rooms";
 
 
 @Component({
@@ -14,44 +16,42 @@ import { selectRoomsAll } from "../../../../store/chat/reducers/rooms";
         <button class="p-0 m-0 leading-none relative left-1" style="top: 1px" title="Compose">
             <i class="fa-regular fa-plus text-xl leading-none text-slate-600" aria-hidden="true"></i>
         </button>
-        <input class="border rounded-lg bg-transparent px-2 w-3/4" placeholder="Search" />
+        <input [(ngModel)]="searchValue" class="border rounded-lg bg-transparent px-2 w-3/4" placeholder="Search" />
     </div>
 
     <div class="flex flex-col gap-2 overflow-y-auto" style="height: calc(100vh - 11rem)">
-        @for (item of (conversations$ | async); track $index) {
-            <app-conv-item [conv]="item" />
+        @for (item of (rooms$ | async); track $index) {
+            <chat-room-item [conv]="item" />
         }
     </div>
 </div>
     `
 })
 export class ConversationsListComponent {
-    conversations$: Observable<UiConv[]>;
+    rooms$: Observable<Room[]>;
+    private searchSubject: BehaviorSubject<string> = new BehaviorSubject<string>("");
 
-    constructor(private store: Store<AppState>) {
-        this.conversations$ = merge(
-            // IMPORTANT: Example data. Will be removed on CU-8695z23hy
-            of(mock_conversations),
-
-            // This is only for rooms, not for 1-to-1 conversations - this
-            // things will be getting from another place
-            this.store.select(selectRoomsAll).pipe(
-                map(data => {
-                    const convs: UiConv[] = data.map(d => ({
-                        id: d.jid as string,
-                        name: d.name as string,
-                        time: new Date(),
-                        preview: "",
-                        unread: 0,
-                        selected: false,
-                        room: true,
-                    }))
-                    return convs
-                })
+    constructor(private store: Store<AppState>, private router: Router) {
+        // select rooms without parent
+        const _root_rooms$ = this.store.select(selectRoomsAll).pipe(
+            map(rooms => rooms
+                .filter(room => room.parent === null)
             ),
-        ).pipe(
-            scan((acc: UiConv[], curr: UiConv[]) => [...acc, ...curr], []),
-            map(convs => convs.sort((a: UiConv, b: UiConv) => b.time.getTime() - a.time.getTime()))
+        );
+
+        // filter if needed by search subject
+        this.rooms$ = combineLatest([_root_rooms$, this.searchSubject]).pipe(
+            map(([rooms]) => rooms
+                .filter(room => room.name.toLowerCase().includes(this.searchValue.toLowerCase()))
+            )
         )
+    }
+
+    get searchValue(): string {
+        return this.searchSubject.getValue()
+    }
+
+    set searchValue(value: string) {
+        this.searchSubject.next(value)
     }
 }