+++ client/projects/client/src/app/modules/chat/components/not-chat-selected.component.ts
@@ -1,0 +1,12 @@
+import { Component } from "@angular/core";
+
+@Component({
+    selector: "no-chat-selected",
+    template: `
+<div class="border rounded-lg p-7 text-slate-800 h-full">
+    No chat selected
+</div>
+`
+})
+export class NoChatSelectedComponent { }
+