+++ client/projects/client/src/app/modules/chat/main-chat.component.ts
@@ -1,13 +1,6 @@
 import { Component } from "@angular/core";
-import { Store } from "@ngrx/store";
 import { UiService } from "@ui";
-import { Observable, Subscription } from "rxjs";
-import { AppState } from "../../store/store";
-import { XmppService } from "../../lib/services/xmpp.service";
-import { XmppEventsDistributionService } from "../../lib/services/xmpp-events-distribution.service";
-import { DiscoItems } from "stanza/protocol";
-import { ChatRoomAdd } from "../../store/chat/actions";
-import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
+import { Observable } from "rxjs";
 
 @Component({
     selector: "app-chat",
@@ -18,25 +11,9 @@ export class MainChatComponent {
 
     constructor(
         private ui: UiService,
-        private store: Store<AppState>,
-        private xmpp: XmppService,
-        private eventsDistribution: XmppEventsDistributionService,
     ) {
         this.ui.title.next("Chat");
         this.loading$ = this.ui.appLoading.asObservable();
-        this.loadRoomsAndDispatch();
-        this.eventsDistribution.run(
-            this.xmpp.receivedMessage$,
-            this.xmpp.receivedPrecense$,
-            this.xmpp.receivedIQ$
-        );
     }
 
-    private loadRoomsAndDispatch() {
-        this.xmpp.queryRoomsOnline().pipe(
-            takeUntilDestroyed(),
-        ).subscribe(data => {
-            (data.disco as DiscoItems).items?.map(d => this.store.dispatch(ChatRoomAdd({ payload: d })));
-        });
-    }
 }