+++ client/projects/client/src/app/store/chat/reducers/participants.ts
@@ -1,34 +1,21 @@
+import { RoomParticipant } from "@lib";
+import { createEntityAdapter, EntityAdapter, EntityState } from "@ngrx/entity";
+import { createReducer, on } from "@ngrx/store";
+import * as actions from "../actions/participants"
+import { fillState, MOCK_CHAT_STORE } from "../mock-store";
 
-import { createReducer, createSelector, on } from "@ngrx/store";
-import { EntityAdapter, EntityState, createEntityAdapter } from "@ngrx/entity";
-import * as actions from '../actions';
-import { ReceivedPresence } from "stanza/protocol";
-import { selectXmppFeature } from "../selectors";
-import { XmppState } from "../reducer";
+export interface ParticipantsState extends EntityState<RoomParticipant> {}
 
-export interface ParticipantsState extends EntityState<ReceivedPresence> { }
-export const ParticipantsAdapter: EntityAdapter<ReceivedPresence> = createEntityAdapter<ReceivedPresence>({
-    selectId: (model: ReceivedPresence) => model.from
-});
-export const initialParticipants: ParticipantsState = ParticipantsAdapter.getInitialState();
+export const ParticipantsAdapter: EntityAdapter<RoomParticipant> = createEntityAdapter<RoomParticipant>({
+    selectId: (model: RoomParticipant) => model.jid
+})
 
-export const participantsReducer = createReducer(
-    initialParticipants,
-    on(actions.ChatParticipantAdd, (state, { payload }) => ParticipantsAdapter.addOne(payload, state)),
-    on(actions.ChatParticipantRemove, (state, { id }) => ParticipantsAdapter.removeOne(id, state)),
+export const ParticipantsInitialState: ParticipantsState = fillState<RoomParticipant>(ParticipantsAdapter, MOCK_CHAT_STORE.participants)
+
+export const ParticipantsReducer = createReducer(
+    ParticipantsInitialState,
+    on(actions.ChatParticipantsAdd, (state, { participant }) => ParticipantsAdapter.addOne(participant, state)),
+    on(actions.ChatParticipantsAddMany, (state, { participants }) => ParticipantsAdapter.addMany(participants, state)),
+    on(actions.ChatParticipantsUpdate, (state, { update }) => ParticipantsAdapter.updateOne(update, state)),
+    on(actions.ChatParticipantsAddRemove, (state, { id }) => ParticipantsAdapter.removeOne(id, state)),
 )
-
-export const selectParticipantsFeature = createSelector(
-    selectXmppFeature,
-    (state: XmppState) => state.participants
-)
-
-export const selectAllParticipants = createSelector(
-    selectParticipantsFeature,
-    ParticipantsAdapter.getSelectors().selectAll
-);
-
-export const selectParticipantsByRoom = (jid: string) => createSelector(
-    selectAllParticipants,
-    (participants: ReceivedPresence[]) => participants.filter(participant => participant.from.startsWith(jid))
-);