+++ client/projects/client/src/app/store/chat/selectors/feature.ts
@@ -1,5 +1,4 @@
 import { createFeatureSelector } from "@ngrx/store";
-import { XmppState } from "./reducer";
+import { XmppState } from "../reducer";
 
-// selectors
 export const selectXmppFeature = createFeatureSelector<XmppState>('chat');