+++ client/projects/client/src/app/store/chat/selectors/users.ts
@@ -1,0 +1,14 @@
+import { createSelector } from "@ngrx/store";
+import { selectXmppFeature } from "./feature";
+import { XmppState } from "../reducer";
+import { UsersAdapter } from "../reducers/users";
+
+export const selectChatUsersFeature = createSelector(
+    selectXmppFeature,
+    (state: XmppState) => state.users
+)
+
+export const selectChatUsersAll = createSelector(
+    selectChatUsersFeature,
+    UsersAdapter.getSelectors().selectAll
+);