+++ client/projects/lib/src/lib/pipes/index.ts
@@ -1,0 +1,2 @@
+export * from "./subroomname.pipe"
+export * from "./namefromjid.pipe"