+++ client/projects/lib/src/lib/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './kratos'
 export * as operations from './xmpp'
+export * from './jid'